backend/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
luno-python==0.0.10
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
urllib3==2.3.0
